# openapi/openapi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "v0/tenant.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v0/tenants/lookup": {
      "get": {
        "summary": "LookupTenantsByEmail returns the active tenants a user belongs to, looked up by email.\nThis endpoint is used by the Login UI before authentication to discover which tenant(s)\nthe user can log in to. Only enabled tenants are returned.\nNOTE: This endpoint is registered unauthenticated on the chi mux (outside the auth router)\nso the Login UI can call it before the user has a token. See ADR 0008.",
        "operationId": "TenantService_LookupTenantsByEmail",
        "responses": {
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "TenantService"
        ]
      }
    }
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
  }
}